database schema change( deleted Grade table

+++ student_management/institution/views/courses.py
@@ -1,4 +1,9 @@
-from student_management.models import Student, Course, Teacher , StudentCourse
+from student_management.models import (
+    Student, Course, 
+    Teacher , 
+    StudentCourse, 
+    Admin , User
+)
 from student_management import db 
 from student_management.student.serializers import students_fields_serializer
 from student_management.utils import random_char
@@ -6,7 +11,7 @@ from flask_restx import Namespace, Resource , fields
 from http import HTTPStatus
 from flask import request
 from ..serializers_course import course_retrieve_fields_serializer
-
+from flask_jwt_extended import jwt_required, get_jwt_identity  
 
 courses_namespace = Namespace('courses', description='Namespace for courses')
 
@@ -18,7 +23,6 @@ course_creation_serializer = courses_namespace.model(
         'teacher_id': fields.Integer(required=True, description="Course teacher id"),
     }
 )
-
 
 student_course_register_serializer = courses_namespace.model(
     'Student Course Creation Serializer', {
@@ -47,6 +51,13 @@ course_retrieve_serializer = courses_namespace.model('Course Retrieval serialize
 class CoursesListView(Resource):
 
     @courses_namespace.marshal_with(course_retrieve_serializer)
+    @courses_namespace.doc(
+        description="""
+            This endpoint is accessible to all users. 
+            It allows the retrieval of all available courses
+            """
+    )
+    @jwt_required()
     def get(self):
         """
         Retrieve all available courses
@@ -54,11 +65,23 @@ class CoursesListView(Resource):
         courses = Course.query.all()
         return courses , HTTPStatus.OK
     
+
     @courses_namespace.expect(course_creation_serializer) 
+    @courses_namespace.doc(
+        description="""
+            This endpoint is accessible to an admin. 
+            It allows admin create a new course
+            """
+    )
+    @jwt_required()
     def post(self):
         """
         Create a new course
         """
+        authenticated_user_id = get_jwt_identity() 
+        admin = Admin.query.filter_by(id=authenticated_user_id).first()   
+        if not admin :
+            return {'message':'You are not authorized to the endpoint'}, HTTPStatus.UNAUTHORIZED
         data = request.get_json()
         teacher = Teacher.query.filter_by(id=data.get('teacher_id')).first()
         if teacher:
@@ -91,12 +114,24 @@ class CourseRetrievalView(Resource):
         course = Course.get_by_id(course_id)
         return course , HTTPStatus.OK
     
-
+    @courses_namespace.doc(
+        description="""
+            This endpoint is accessible to an admin. 
+            It allows admin delete a course
+            """
+    )
+    @jwt_required()
     def delete(self, course_id):
         """
         Delete a course
         """
-        course = Course.get_by_id(course_id)
+        authenticated_user_id = get_jwt_identity() 
+        admin = Admin.query.filter_by(id=authenticated_user_id).first()   
+        if not admin :
+            return {'message':'You are not authorized to the endpoint'}, HTTPStatus.UNAUTHORIZED
+        course = Course.query.filter_by(course_id).first()
+        if not course:
+            return {'message':'Course does not exist'}, HTTPStatus.NOT_FOUND
         try:
             course.delete()
         except:
@@ -106,18 +141,33 @@ class CourseRetrievalView(Resource):
     
 
 
-@courses_namespace.route('/<int:course_id>/add_and_drop')
+@courses_namespace.route('/<int:course_id>/students/add_and_drop')
 class CourseRetrievalView(Resource):
 
+
     @courses_namespace.expect(student_course_register_serializer)
+    @courses_namespace.doc(
+        description="""
+            This endpoint is accessible to a teacher. 
+            It allows teacher add a  student the their course
+            """
+    )
+    @jwt_required()
     def post(self, course_id ):
         """
         Register a student to a course
         """
-        course = Course.get_by_id(course_id)
+        authenticated_user_id = get_jwt_identity() 
+        teacher = Teacher.query.filter_by(id=authenticated_user_id).first()   
+        if not teacher :
+            return {'message':'You are not authorized to the endpoint'}, HTTPStatus.UNAUTHORIZED
         data = request.get_json()
         student_id = data.get('student_id')
+        # check if student and course exist
         student = Student.query.filter_by(id=student_id).first()
+        course = Course.query.filter_by(course_id).first()
+        if not student or not course:
+            return {'message': 'Student or course not found'}, HTTPStatus.NOT_FOUND
         if student:
             #check if student has registered for the course before
             get_student_in_course = StudentCourse.query.filter_by(student_id=student.id, course_id=course.id).first()
@@ -136,15 +186,28 @@ class CourseRetrievalView(Resource):
         return {'message': 'Student does not exist'} , HTTPStatus.NOT_FOUND
     
 
-
+    @courses_namespace.doc(
+        description="""
+            This endpoint is accessible to a teacher. 
+            It allows teacher remove a  student from their course
+            """
+    )
+    @jwt_required()
     def delete(self, course_id):
         """
         Unregister a student course
         """
-        course = Course.get_by_id(course_id)
+        authenticated_user_id = get_jwt_identity() 
+        teacher = Teacher.query.filter_by(id=authenticated_user_id).first()   
+        if not teacher :
+            return {'message':'You are not authorized to the endpoint'}, HTTPStatus.UNAUTHORIZED
         data = request.get_json()
         student_id = data.get('student_id')
+        # check if student and course exist
         student = Student.query.filter_by(id=student_id).first()
+        course = Course.query.filter_by(course_id).first()
+        if not student or not course:
+            return {'message': 'Student or course not found'}, HTTPStatus.NOT_FOUND
         if student:
             #check if student has registered for the course before
             get_student_in_course = StudentCourse.query.filter_by(student_id=student.id, course_id=course.id).first()
@@ -167,10 +230,25 @@ class CourseRetrievalView(Resource):
 class CourseStudentsListView(Resource):
 
     @courses_namespace.marshal_with(students_serializer) 
+    @courses_namespace.doc(
+        description="""
+            This endpoint is accessible to an admin and a teacher. 
+            It allows the retrieval of all students in a course
+            """
+    )
+    @jwt_required()
     def get(self, course_id ):
         """
         Retrieve all registered student in a course
         """
+        authenticated_user_id = get_jwt_identity() 
+        user = User.query.filter_by(id=authenticated_user_id).first()   
+        if not user or user.user_type == 'student' :
+            return {
+                'message':'You are not authorized to the endpoint'
+                }, HTTPStatus.UNAUTHORIZED
+        
+        
         course = Course.get_by_id(course_id)
         get_student_in_course = StudentCourse.get_students_in_course_by(course.id) 
         return get_student_in_course , HTTPStatus.OK
